+++ CHANGELOG.md
@@ -1,4 +1,23 @@
-### 8.0.0
+## 8.0.2
+
+##### Bugfix
+
+-   Add `fixAndroidTouchableBug` prop (#354)
+
+---
+
+## 8.0.1
+
+##### Bugfix
+
+-   Locked @react-native-community/picker to 1.6.0 to fix Expo issues
+-   Add togglePicker method to Picker component typescript definition (#360)
+-   Fix wrong PickerProps import in index.d.ts (#352)
+-   Fixed inputWeb to be a TextStyle, not ViewStyle (#365)
+
+---
+
+## 8.0.0
 
 ##### Breaking Changes
 
@@ -11,7 +30,7 @@
 
 ---
 
-### 7.0.0
+## 7.0.0
 
 ##### Breaking Changes